requirements.txt
meshtastic==2.7.1
pyserial==3.5
pydantic==2.11.7
customtkinter==5.2.2
CTkToolTip==0.8
keyring==25.6.0